Disable glance-registry
As described in [1], [2], glance-registry was deprecated from Queens and will be
removed in S, so we should disable glance registry in Queens but keep the image
and deploy logic if user need this (but it's low chance due to glance-registry
almost do nothing currently).

This patch only disable glance registry microservice.

[1]: http://specs.openstack.org/openstack/glance-specs/specs/newton/approved/glance/deprecate-registry.html
[2]: https://blueprints.launchpad.net/glance/+spec/deprecate-registry

Change-Id: I6667325711d1c54cf5233ad6973d89ea396a46b5
Closes-bug: #1769837