Avoid running kolla source jobs for unnecessary  files
While fixing test-requirement.txt, I found that
check pipeline can be more optimized in term of selecting
the jobs to run based on file change.
- https://review.openstack.org/#/c/567083/
kolla main jobs are not needed to be run on check
pipeline for test-requirements.txt, releasenotes/* etc.

This commit skip those jobs when only above files
are changed in patch. This is will save time as well
as infra resource also.

Change-Id: I0c1328237d2109b1d1d5dcb40f331fff1ab10257