Fix default value logic of 'glance_backend_file' variable
glance_backend_file should be set to 'yes' if no other backends
are enabled. In all-Ceph clusters the best choice of backend is
Ceph. In non-Ceph clusters if Swift is enabled the latter is the
next best choice for Glance. With this fix default backend is
configured automatically and helps avoid non-ha file backend
issues after deployment.

Change-Id: I3d1c7cb2de9fedecdf9ad8744f6589c6f3f83ece
Closes-Bug: #1770660