Avoid removing kolla-cli directories on destroy
When running the kolla-ansible destroy playbook most of the files
under /etc/kolla are cleaned up.  This change will cause the
clean-up to skip the kolla-cli related directories.  When running
the deployment code in a container or a separate VM this is not
an issue, but when doing an all-in-one or running the deployment
process from one of the systems being deployed onto it would have
previously destroyed the kolla-cli inventory and any generated
certificates used for TLS.

Change-Id: Icd62e177964b8fb31052131785c6d7b138975a28