Update install.yml
I kept running into an issue after upgrading to pip 10 that the docker SDK couldn't get installed due to the error below.  So I added the extra args that doesn't care and ignores the install packages.

Cannot uninstall 'requests'. It is a distutils installed project and thus we cannot accurately determine which files belong to it which would lead to only a partial uninstall.