Create /volumes when using external ceph
Currently we only create the /volumes directory when
using the kolla-ansible bundled ceph. This extends the
functionality to external ceph.

Note: In addition to the cherry-picked commit, we add jmespath to
requirements as it is required for the json_query filter, with a minimum
version of 0.9.3 to match lower-constraints.txt.

Change-Id: I1e9a40e4ee893a53f04909ef7835be005c9b5276
Closes-Bug: #1778720
(cherry picked from commit 2f6b1c6890cf7ea6b0dd33ac219646e4dcaf1fd6)