Fix etcd when Kuryr is disabled
This is because the environment variable ETCD_INITIAL_CLUSTER contains
compute nodes, which are not running etcd.

Compute nodes are in the etcd group to allow for kuryr, which needs etcd
to be running on compute nodes (why???). When kuryr is disabled, etcd is
not deployed on these hosts (using host_in_groups). However, the
ETCD_INITIAL_CLUSTER environment variable always includes all hosts in
the etcd group.

This change sets ETCD_INITIAL_CLUSTER correctly in both cases.

Change-Id: I11f812d3fdacf0eeb8236f9a39d673aaa480dd26
Closes-Bug: #1783963