Run etcd on controller node only
In before, etcd container is launched in compute node if kuryr is
enabled. This logic breaks the all-in-one setup in which the etcd
cluster is configured with two members (controller and compute)
but it actually has one member only.

However, even if kuryr is enabled, running etcd in compute node
is not the common practice and kuryr has changed its devstack
plugin to run etcd on controller only [1].

[1] https://review.openstack.org/#/c/514815/

Closes-Bug: #1790415
Change-Id: I701d495675178c3ed8ec1f00b31d09f198b38a6f