Fix on setting enable_openvswitch
In the default expression which sets enable_openvswitch, the boolean filter
is applied only on 'linuxbridge', not on the whole comparing expression. So
even if neutron_plugin_agent is set to 'linuxbridge', but enable_openvswitch
is not explicitly set to "no" (e.g., in globals.yml), then enable_openvswitch
is eventually set to true, and openvswitch role is unintentionally applied.
Boolean filter is not really necessary on the comparing expression, so it can
be removed just like in globals.yml.

Change-Id: I0612346484eada8192ee03dc9d5057cb6881f532
Closes-bug: #1747136