Allow gnocchi services to use independent hostnames
This allows gnocchi service endpoints to use custom hostnames, and adds the
following variables:

* gnocchi_internal_fqdn
* gnocchi_external_fqdn

These default to the old values of kolla_internal_fqdn or
kolla_external_fqdn.

This also adds a gnocchi_api_listen_port option, which defaults to
gnocchi_api_port for backward compatibility.

This option allow the user to differentiate between the port the
service listens on, and the port the service is reachable on. This is
useful for external load balancers which live on the same host as the
service itself.

Change-Id: Ic9a0f8130b19ed77987f45fd0e824b82ea7a7328
Implements: blueprint service-hostnames