Hide logs when looping over passwords
When ansible goes in to a loop, by default it prints all the keys for
the item it is looping over. Some roles, when setting up the databases,
iterate over an object that includes the database password.

Override the loop label to hide everything but the database name.

Change-Id: I336a81a5ecd824ace7d40e9a35942a1c853554cd
(cherry picked from commit 30c619d1bc8938ea2177fdc21e503808dbe0cca7)