Automate Monasca documentation for configuring Kafka
Until the Monasca Kafka client fork is removed it is currently required
to run Kafka in compatibility mode. It is also necessary to disable
an optimisation in the Kafka brokers to clean up idle connections. This
is because the optimisation was added after the Monasca Kafka client was
forked, and the client hasn't been updated since. These settings are now
applied automatically when Monasca is enabled.

Change-Id: I6935f1fb29f4f731cf3c9a70a0adf4d5812ca55e