Make Kafka maintenance easier
This commit enables two settings which are useful when restarting a Kafka node
as part of a cluster. The first supports moving partitions for which a node
being restarting is the leader to another broker, and the second supports
automatically rebalancing partitions when the node rejoins the cluster. See
the documentation for more details:

https://kafka.apache.org/10/documentation.html#basic_ops_restarting

Change-Id: I0718fca08ea116db926b348e3e6e3bc2373db1fb